requirements.txt
Flask==2.1.0
Flask-CORS==3.0.10
requests==2.26.0
wikipedia==1.4.0
torch
scipy
Werkzeug==2.2.2
nltk
openai
python-dotenv
styletts2
numpy
llama-cpp-python